backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.0
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0
python-multipart==0.0.6
nltk==3.8.1
lxml==4.9.3
